server/requirements.txt
Flask==0.12
pyserial==3.4